improve sales module ui in Boilerplate #9807
