+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/Components/Layout/IdentityHeader.razor
@@ -37,7 +37,7 @@
                    Variant="BitVariant.Text"
                    OnClick="WrapHandled(ToggleTheme)"
                    Color="BitColor.SecondaryBackground"
-                   IconName="@(currentTheme == AppThemeType.Light ? BitIconName.Sunny : BitIconName.ClearNight)" />
+                   IconName="@(currentTheme == AppThemeType.Light ? BitIconName.ClearNight : BitIconName.Sunny)" />
 
         @if (CultureInfoManager.MultilingualEnabled)
         {
