+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/Components/Layout/MainLayout.razor
@@ -7,7 +7,7 @@
             <CascadingValue Name=@Parameters.IsAuthenticated Value=isAuthenticated>
                 <CascadingValue Name=@Parameters.CurrentRouteData Value=currentRouteData>
                     <CascadingValue Name=@Parameters.IsCrossLayoutPage Value=isCrossLayoutPage>
-                        <BitAppShell>
+                        <BitAppShell AutoGoToTop>
                             <main class="@GetMainCssClass()">
                                 <div class="root">
                                     @if (isAuthenticated is false)
