+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/Components/Pages/ProductImage.razor
@@ -1,4 +1,4 @@
-<BitImage Src="@(Src ?? "_content/Boilerplate.Client.Core/images/product-placeholder.png")" Width="@Width" ShouldFadeIn="false" />
+<BitImage Src="@(Src ?? "_content/Boilerplate.Client.Core/images/product-placeholder.svg")" Width="@Width" ShouldFadeIn="false" />
 
 @code {
     [Parameter] public string? Src { get; set; }
