+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/compilerconfig.json
@@ -73,8 +73,8 @@
         "options": { "sourceMap": false }
     },
     {
-        "outputFile": "Components/Pages/HomePage.razor.css",
-        "inputFile": "Components/Pages/HomePage.razor.scss",
+        "outputFile": "Components/Pages/Home/HomePage.razor.css",
+        "inputFile": "Components/Pages/Home/HomePage.razor.scss",
         "minify": { "enabled": false },
         "options": { "sourceMap": false }
     },
